add cast to float

+++ src/java/org/apache/cassandra/config/Config.java
@@ -93,7 +93,7 @@ public class Config
     public Integer memtable_flush_writers = null;
     public Integer memtable_heap_space_in_mb;
     public Integer memtable_offheap_space_in_mb;
-    public Double memtable_cleanup_threshold = null;
+    public Float memtable_cleanup_threshold = null;
 
     public Integer storage_port = 7000;
     public Integer ssl_storage_port = 7001;
